+++ actix-cors/CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Use `std::future::{ready, Ready}` instead of `actix-utils` types.
+
 ## 0.7.2
 
 - Add configured CORS response headers to error responses.